BREAKING CHANGE: change `Storable` encoding of integers from le to be (#14)
Update the way integers are serialized/deserialized to use big-endian
bytes rather than little-endian. That way, the byte representations of
these numbers are sorted, which is a very desirable property.

I asked all the current users of `stable-structures` and none are using
these encodings yet, so this should be a safe breaking change.