chore: Enhancements to the StableBTreeMap API. (#17)
BTreeMap currently take the maximum size of the BTreeMap as arguments.
Example:

```
let map: StableBTreeMap<_, Key, Value> = StableBTreeMap::init(memory, MAX_KEY_SIZE, MAX_VALUE_SIZE)
```

One issue with this API is that specifying the max sizes is independent
of the type being used, and that poses a couple of usability issues:

1. If you're using a primitive type as your key/value (e.g. `u64`), it's
not obvious what size you should put as an argument. The correct answer
is `8`, because the `Storable` implementation of `u64` serializes it to
a vector of 8 bytes, but that's far from obvious.
2. It's easy to put an incorrect value by accident if you're managing a
few stable structures. I already made this mistake in the bitcoin
canister.

A more safe and clear implementation is that the `Key` and `Value`
contain information about their size, so that users don't need to
specify the size explicitly when initializing the map.

This commit introduces a `BoundedStorable` trait, where it now required
for a type to be used in `StableBTreeMap`. The map is then able to fetch
the maximum size of that type without it being passed explicitly.

```
impl BoundedStorable for Key {
  fn max_size() -> u32 {
     MAX_KEY_SIZE
  }
}

impl BoundedStorable for Value {
  fn max_size() -> u32 {
     MAX_VALUE_SIZE
  }
}
```

And now the syntax is closer to a standard `BTreeMap`:

```
let map: StableBTreeMap<_, Key, Value> = StableBTreeMap::init(memory)
```

Storing the maximum size in the type allows us in the future to also
support blanket implementations of tuples, so one can then implement:

```
let map: StableBTreeMap<_, (A, B), (C, D)> = StableBTreeMap::init(memory)
```


The updated examples showcase how the new API is used.