fix: do not make `BoundedStorable` fixed size by default (#30)
It's clearer to make developers explicitly specify whether or not a type
is fixed or variable in size.

With new optimizations that we'll be doing to `BTreeMap`, variable sized
elements can end up having a smaller memory footprint. If we assume by
default that a type is fixed in size, developers won't benefit from this
optimization.