feat(btreemap): implement proper range scan
This change replaces the `range` method based on prefixes with a
`range` method based on key range, similarly to the `BTreeMap::range`
from the standard library.

To achieve this, we:

  1. Require map keys to implement Ord.
  2. Store node keys decoded.
  3. Store the key bounds in the map iterator.

I also separated keys and values into separate arrays, because it
makes the code a bit cleaner and usually a bit more efficient.