fix(memorymanager): remove unneeded assertion (#44)
Problem:
When loading a `MemoryManager`, the code asserts that the bucket size is
equal to the `BUCKET_SIZE_IN_PAGES` constant. In practice, this makes
upgrading to 0.4.0 from previous versions fail, given that the
`BUCKET_SIZE_IN_PAGES` has been changed.

Solution:
Remove the assertion, as it is no longer required.