fix(log, vec): use u64 for indices and lengths
On the Wasm32 architecture, usize is 32 bit, making it impossible for
the log to report correct size in bytes on large data sets.

This commit changes the types of indices and lengths in Vec and Log so
that they will behave correctly if these structures grow beyond four
billion items.