feat: add benchmarks for BTreeMap (#50)
# Problem
The benchmarks we have aren't comprehensive enough to get an accurate
picture of the performance of `BTreeMap`. For example, in the `0.4.0`
release, the benchmarking numbers showed us an improvement of ~40% in
the `insert` and `remove` operations, but when I profiled the Bitcoin
canister with that upgrade I saw a _regression_ of about 40%.

# Solution
Add more thorough benchmarks for `insert`, `remove`, and `get` to get a
more accurate picture of the performance impact of our changes. These
will then be used to address the performance regressions we've seen with
`0.4.0+`