Run benchmarks with Criterion
When adding benchmarks that may take a longer time to run or require
setting up large canisters, it would be helpful to use a bench harness
that allows easily selecting which benches to run. This MR does that by
running the benchmarks in Criterion. This also has the added benefit of
being able to quickly compare the results of two runs since the output
shows a percentile improvement/regression.