feat(btreemap): add iter_upper_bound api
This change introduces a new btree map method, `iter_upper_bound`. It
returns an iterator pointing to the first element that is less than the
given key.

We need this method for the time series API because we need to find the
closest keys to the left and to the right of the given key. We can do
the latter with the `range` API, but the former requires a new method.

We took inspiration from experimental API in BTreeMap:
https://doc.rust-lang.org/nightly/std/collections/struct.BTreeMap.html#method.upper_bound