refactor: generalize the `Node` in `BTreeMap` (#74)
This commit lays some of the groundwork for EXC-1395, which is to
support variable-length keys and values in the BTreeMap.

The code that is specific to the node containing a fixed number of
elements is abstracted away from the `BTreeMap` and into the `Node`
itself. Once the concept of a node is completely abstracted, then we can
drop in a new implementation of Node that supports a variable number of
entries.

As part of this abstraction:
* Some of the fields/constants in `node.rs` are made private. Follow-up
commits will make all fields in the `Node` private.
* Some tests were rewritten to be more generic, but still offers the
same or better coverage.