fix: use drun for running the benchmarks. (#77)
# Problem
Benchmarks are run within a canister to obtain the instruction count.
However, benchmarks can run for up to 5B instructions in case of
queries, and 50B in case of updates. We do have benchmarks that we
haven't been able to run because they surpass these limits.

# Solution
Use `drun` instead of `dfx` to run the benchmarks. `drun` can be
configured to have a practically infinite instruction count. As a
result, a few benchmarks are now enabled by this commit, and all the
benchmarks are now query calls to avoid side effects.

Closes EXC-1405