feat: v2 of `BTreeMap::Node` to support unbounded types. (#114)
Introduces V2 of `BTreeMap::Node`, which has a smaller memory footprint
than V1 and includes support for unbounded types.

The implementation is currently not very efficient, as it deserializes
all entries on load and serializes all entries on save. Performance
enhancements will be included in subsequent PRs.

Note that V2 here is only used in tests and production continues to use
V1.