feat: introduce new benchmarking script (#121)
## Problem
We've been relying on `criterion` as our way of running benchmarks and
getting reports. However, `criterion` has been falling short for a
number of reasons:

1. Benchmarks can only have one measurement. In our case, the
measurement was "instructions", but that's not good enough. For our
use-case, we'd like to have multiple measurements per benchmark (e.g.
instructions, stable memory pages, etc.).
2. `criterion` always compares the results to the previous run. In
almost all cases, I'm interested in measuring the difference in
performance relative to the last commit, and there isn't a good of
enforcing that workflow in `criteron`.

## Solution
I decided to spend a few hours to replace our reliance on `criterion`
altogether. We now have our very own benchmarking script that runs the
benchmarks. A benchmark can have as many measurements as we want, and it
neatly persists these benchmarks to our repo along with a report on how
the benchmarks have changed.

Here's an example of the report that it prints out to the terminal:
<img width="435" alt="image"
src="https://github.com/dfinity/stable-structures/assets/208628/0d5efd72-fe76-4246-b68b-e9bbe7c20102">

Notes:
* As you can tell from the screen shot above, I added the
"stable_memory_size" as a new measurement.
* I had to bump up the rust version as `clap` needed a more rust >=
1.70.