refactor: determine node version automatically on load (#133)
Prior to this commit, the caller of `Node::load` needed to specify the
node version that will be loaded. This is acceptable if all nodes have
the same version, but given that a BTreeMap may have different versions
of nodes (migration will be introduced in subsequent commits), this
needs to change such that the node's version is detected on load.