feat: add additional assertions to the `Storable` trait. (#136)
This commit adds a new method, to_bytes_checked, to the Storable trait.
It includes assertions to detect if the Storable implementation of an
element is buggy. This addresses the issue in
https://github.com/dfinity/stable-structures/pull/125, where a user had
a mistake in their implementation of Storable, but it wasn't clear from
the error that this was the case.

BTreeMap has been updated to use to_bytes_checked rather than to_bytes.