feat: make BTreeMap V2 the default and automatically migrate V1 maps. (#135)
Prior to this commit, initializing a BTreeMap initialized a V1 BTreeMap
by default, and only by using `new_v2` and `init_v2` can developers
explicitly opt-in to using V2 BTreeMaps.

This commit updates the interface such that a BTreeMap is initialized as
a V2 BTreeMap by default, and v1 BTreeMaps are automatically migrated to
V2. Developers can still opt-in to using V1 by using `new_v1` and
`init_v1`.