feat: detect benchmarks automatically.
Problem: In the benchmarking script, we maintain a hard-coded list of
all the benchmarks to run. These can get out of date, as currently was
the case, causing some benchmarks to never run.

Solution: All the benchmarks are queries. Rather than maintaining a
hard-coded list, we now parse the Wasm to identify these query methods
directly.

NOTE: Some benchmarks were not being run, and have now been added to
the `results.yml` file.