fix: BTreeMap memory leak when deallocating nodes with overflows (#212)
# Problem
When storing unbounded types in a `BTreeMap`, a node is represented as a
linked list of "memory chunks". It was discovered recently that when we
deallocate a node, in some cases only the first memory chunk is
deallocated, and the rest of the memory chunks remain (incorrectly)
allocated, causing a memory leak.

# Solution
Change the logic for deallocating nodes to ensure that all of a node's
memory chunks are deallocated. Tests have been added to prevent
regressions of this nature moving forward.