perf: Add new efficient APIs read_unsafe and read_to_vec (#248)
I found that a source of significant performance loss is the read method
of `Memory`. The `read` method takes a mutable buffer which it fills
with values read from the stable memory. According to Rust rules, the
buffer passed to read must be initialized before it's passed to read
(buffers containing uninitialized values are unsound and can cause UB).

The usual pattern is to create a properly sized Vec, eg. by using
`vec![0; size]` or `vec.resize(size, 0)` and pass that to `read`.
However, initializing the bytes with values that get overwritten by
`read` is only necessary in order to be sound and requires significant
number of instructions.

This PR introduces a new method `read_unsafe` which allows passing in a
raw pointer and a `count` parameter. Implementations can be more
efficient by reading directly and skipping initialization. This can lead
to instruction reductions of up to 40%.

The PR also introduces a helper method `read_to_vec` which is a safe
wrapper around `read_unsafe` for the most common use-case: reading into
a `Vec`. Clients can for example pass an empty `Vec` and profit from the
extra efficiency without having to call unsafe methods.

---------

Co-authored-by: Andriy Berestovskyy <berestovskyy@gmail.com>