perf: Replace BucketIterator with function (#258)
Iterating through the buckets with `BucketIterator` was nice because we
could use a for loop but the compiler cannot optimize it as well as a
function that takes a callback. Since this code has to be run on every
single memory read/write, making it as efficient as possible is crucial.
Using a function not only allows for better wasm generation but we can
further optimize the implementation more easily (e.g. we can add a
TLB-like cache).