docs: add skeleton for stable structures book (#289)
There are a number of concepts in stable structures that need more
thorough documentation, and I think it warrants some kind of "book" or
reference guide that can be used by developers.

This commit adds the skeleton of what would become this book, and in the
future we'll use mdbook to serve this documentation.

An example of what this book will look like:
https://proptest-rs.github.io/proptest/